backend/requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6
django-filter>=23.5,<24
django-cors-headers>=4.3,<5
psycopg[binary]>=3.1,<4
gunicorn>=21.0,<23
sentry-sdk[django]>=1.39,<3
